pyproject.toml
[build-system]
requires = ["setuptools >= 61", "setuptools_scm[toml]>=3.4", "wheel"]

[project]
name = "biahub"
description = "Bioimage analysis hub."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11, <4.0"

# the dynamically determined project metadata attributes
dynamic = ["version"]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
]

# list package dependencies here
dependencies = [
  "iohub[tensorstore] @ git+https://github.com/czbiohub-sf/iohub.git@xarray-integration",
  "stitch @ git+https://github.com/ahillsley/stitching@jen",
  "matplotlib",
  "natsort",
  "numpy",
  "markdown",
  "monai",
  "pandas~=2.1",
  "pydantic>2,<3",
  "scipy<1.16",  # as of 06/23/2025 scipy 1.16.0 breaks test imports
  "submitit",
  "torch",
  "tqdm",
  "waveorder @ git+https://github.com/mehta-lab/waveorder.git",
  "largestinteriorrectangle",
  "antspyx",
  "pystackreg",
  "imageio-ffmpeg",
  "dask[array]",
  "toml",
  "scikit-learn",
  "numba>=0.58.0",
  "llvmlite>=0.41.0",
  "humanize",
]

[project.optional-dependencies]
segment = ["cellpose"]

track = ["ultrack>=0.7.0rc2"]

gui = ["napari", "PyQt6", "napari-animation"]

shard = ["tensorstore"]

build = ["build", "twine"]

all = ["biahub[segment,track,shard,build,gui]"]

dev = [
  "biahub[segment,track,shard,build]",
  "black==25.1",
  "flake8==7.2",
  "isort==6.0",
  "pytest==8.4",
  "pre-commit~=4.2",
]
doc = [
    "zensical>=0.0.11",
    "mkdocs-click>=0.8",
    "mkdocstrings-python>=2.0.1",
]

[project.scripts]
biahub = "biahub.cli.main:cli"

[tool.setuptools]
# explicitly specify the package name
# TODO: is there a better way to do this?
packages = ["biahub"]

# allow use of __file__ to load data files included in the package
zip-safe = false

[tool.setuptools.dynamic]
version = { attr = "biahub.__version__" }

[tool.black]
line-length = 95
target-version = ['py310']
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | notebooks
    | ignore
    | examples
    | scripts
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 95
lines_between_types = 1
default_section = "THIRDPARTY"
no_lines_before = ["STDLIB"]
ensure_newline_before_comments = true
skip_glob = ["examples/*", "scripts/*"]

[tool.pylint]
# disable all conventions, refactors, warnings (C, R, W) and the following:
# E0401: unable-to-import (since it is possible that no one environment has all required packages)
# E1136: unsubscriptable-object (anecdotal false positives for numpy objects)
disable = ["C", "R", "W", "import-error", "unsubscriptable-object"]
msg-template = "{line},{column},{category},{symbol}:{msg}"
reports = "n"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]
